Fix supervisor shutdown (#9)
Detach the children from the process group so they dont get the SIGINT
signal when pressing ctrl-c.

Add a "quit" channel to indicate that the supervised process should be
shut down. Catch this and send the SIGTERM from the supervisor loop.
This way we avoid sending a SIGTERM to a non-existing process in case
the shutdown happens during respawn pause.

Send a message back to the Stop() function via a "done" channel so we
block til process is terminated.

Stop services in reverse start order.

Signed-off-by: Natanael Copa <ncopa@mirantis.com>