Prevent unnecessary docker exports
tar will set the timestamp of the binary file to be the same as it was
during compile time, which is older than the .container.* stamp. So make
will always extract the binary. Prevent this to happen by touching it
after extraction.

We also add make targets for the intermediate files so they dont get
deleted. See
https://www.gnu.org/software/make/manual/html_node/Chained-Rules.html

Signed-off-by: Natanael Copa <ncopa@mirantis.com>