Use true leader elector for controller counting
Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>

The previous iteration, while successfull at proving the approach in general, had a drawback of the konnectivity-server process flapping bit too much.
This has the unwelcome side-effect of the agents getting slightly confused to which and how many servers they should connect with. This causes connectivity issues between API and workers as all that comms is done through konnectivity tunnels.

This commit changes couple things:
- standard leader election used for per controller leases. They are more accurate but do use bit more resources.
- Instead of restarting the whole konnectivity component we now only restart the supervisor part (== the k-server process itself)
- per controller lease broken into separate component, to keep things bit apart

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>