Store kubelet config as a template in ConfigMap
We need to set Client CA path and volume plugin dir relative the data
dir, and it needs to be set on the worker side. This means that we
cannot save the full path in the ConfigMap. Instead we store a template
in ConfigMap which is rendered on the worker side before writing the
config.

Based on work of @jnummelin

Fixes #771

Signed-off-by: Natanael Copa <ncopa@mirantis.com>