Move `k0s-cloud-provider` logic to Annotations
The initial implementation of `k0s-cloud-provider` relied on querying
`k0sproject.io/node-ip-external` node labels.  While this was fine
for IPv4 addresses, the character restriction for labels prevented
the use of IPv6 addresses.

Instead of using labels, use annotations instead which are not as
restrictive as labels (and should have been used in the first place)

Going forward, the `k0sproject.io/node-ip-external` annotation should
be used for both IPv4/6.

Fixes: #996

Signed-off-by: Shane Jarych <sjarych@mirantis.com>