Make component reconcilers "reactive" instead of polling the config CR
To avoid all components polling for the config CR this commit changes the architecture of reconciliation to be reactive.
The main component "watching" the config CR will trigger call to Reconcile() function of all components. The Reconcile is separated into it's
own interface thus not all components need to implement it and only the reconcilable components are triggered reactively.

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>