Mention that git tags should be annotated in release docs
Annotated tags will make the release date in github to use the date of
the tag. Lightweight tags will use the date of the commit the tag points
to. Using annotated tags makes it possible to set the date of the tag so
correct tag is shown as "latest" in github.

Add a link to git documentation that explains what an annotated tag is.
This will also highlight `annotated` so it stands out better.

Signed-off-by: Natanael Copa <ncopa@mirantis.com>