Make reset smoke test wait till all pods are running
When running reset at the time when kubelet and containerd are still creating lot of the containers the reset might not always be able to remove all containers as they might be created during the removing of existing containers.
This made the smoke test somewhat flaky which this commit hopefully mitigates.

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>