Use k8s version from embedded-bins for conformance tests
We can not simply strip out the `+k0s.*` from the release tag to
calculate the kubernetes version. This is because we use version from
both kuberentes and k0s so when we tag k0s release candidates we set -rc
suffix to the kubernets version part of the string, not to the +k0s
part, because that is not valid semver.

To avoid confusion of kubernetes version release candidates and k0s
version release candidate we do not use the release tag at all to find
kubernetes version. Instead we re-use the kubernetes_version set in
embedded-bins/Makefile.variables, which always will be correct.

While at it, we also remove a hardcoded K8S_VERSION for the network
conformance test.

Fixes commit 6ad33ed56426 (Add CNCF conformance test)

Signed-off-by: Natanael Copa <ncopa@mirantis.com>