Fix storage for --single regression
Kine should be used as default storage with --single option specified.
commit 22ffa9fb80c7 (ClusterConfig CRD: remove k0Vars in favor of
DataDir var) introduced a regression that caused etcd be used by default
with --single.

Fix this and refactor so that DefaultClusterConfig takes an
optional/variadic clusterConfig. This way we don't always need to
specify which default cluster spec is to be used. Then we set the kine
storage in `ValidateYaml` depending on the DefaultStorageType in
K0sVars. `ValidateYaml` needs data dir from k0s vars so it makes sense
to do this from here.

Fixes https://github.com/k0sproject/k0s/issues/1360

Signed-off-by: Natanael Copa <ncopa@mirantis.com>