Explicitly set permissions on created directories
Set permissions to what we want instead of error out when they are not
what we expect them to be.

Umask may be stricter than expected and would lead to error with wrong
message on creation:  "directory ... exist, but the permission is ...."

Only verifying the permission and error out does not provide much value,
so we simply set the permission to what it should be. This makes it also
possible to let have future upgrades with different permissions without
erroring out.

Fixes https://github.com/k0sproject/k0s/issues/587

Signed-off-by: Natanael Copa <ncopa@mirantis.com>