remove --config as a global flag
this commit removes config as a global flag so that it will not
show up in commands that don't use it (for example, worker).
it also sets the default config path to /etc/k0s/k0s.yaml since it is
already the de-facto location for k0sctl and also appears in our
documenttation.
Signed-off-by: Karen Almog <kalmog@mirantis.com>