Ignore chmod events in debouncer
The debouncer is only used the stack applier. We don't need to apply
anything if the permission changes so make the debouncer ignore those
events.

This also makes the config change integration test to pas, which for
some reason fails since commit 6fcd992459cc (Explicitly set permissions
on created directories).

Signed-off-by: Natanael Copa <ncopa@mirantis.com>