Add an 'id' to plan command status
Adding an identifier to plan command status fixes a problem in signal node
reconciliation where multiple commands of the same name (ie. k0supdate + k0supdate)
wouldn't be possible.

This also removes the redundant 'CommandUpdateItem' struct in favor for embedding
the 'k0s' and 'airgap' update structs into the 'Command' struct. This aligns better
with the CRD structures.

This is a breaking change.

Fixes #227

Signed-off-by: Shane Jarych <sjarych@mirantis.com>