Add a new `PlanStateHandler` for plan command initialization.
This addresses a problem in the original multi-command implementation where
the status of all commands in a plan were not initialized properly. In order
for the plan controller to know when the plan is "Completed", it needs to
know the status of every command -- if the status of all commands are not
present, its a problem.

Fixes #221

Signed-off-by: Shane Jarych <sjarych@mirantis.com>