Fix go mod tidy
After the upgrade to cfssl 1.6.1, k0s pulls in etcd 3.4 packages, which
have known issues when used as go modules. This results in breaking go
mod tidy.

Fix this by updating the transitive dependency that pulled in the
problematic etcd package versions to a version in which it pulls version
3.5 instead, which is working quite well.

The command issued was:

  go get github.com/google/certificate-transparency-go@80b9f2a11acd18d19fac9c341532bc8c06ded6aa

Also add a minimalist consistency check to CI to uncover such problems
earlier in the future.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>