Add component to provision static pod manifests
The static_pods component starts a local HTTP server to serve static pod
manifests. The HTTP endpoint is configured as `staticPodURL` in the
Kubelet. This enables k0s to manage node-local pods, without the need of
a connection to the API server. Note that this now "claims" the
`staticPodURL` Kubelet config for internal use inside k0s.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>