Fix endless recursion in CallParentPersistentPreRun
The cobra code calls PersistentPreRun of parents with the actual
command, not with the parent command. CallParentPersistentPreRun did the
same, but that led to endless recursion, since it would find the same
method over and over again, when walking up the command stack. Fix this
by swapping out the pre-run methods temporarily.

Also:
* Exclude tests from k0s's build dependencies in Makefile

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>