Remove unused Reconcile() methods from components
This is probably a leftover from a previous refactoring. Those methods
are not matching the signatures as required by ReconcilerComponent, and
weren't doing anything useful in the first place. Also add a constraint
to all the structs to actually adhere to the component interface and
make some structs public in order to satisfy the linker.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>