Restart leader election as long as k0s is running
The client-go leader elector will return from its run loop whenever the
"leader election loop is stopped by ctx or it has stopped holding the
leader lease". This means that k0s needs to restart leader election as
long as it's running, because the run loop may terminate even if the
context is not yet done.

Implement a simple loop over the leader elector's run method that
terminates as soon as the context is done. Add a test case that
demonstrates a case when leader election will simply stop, even if the
lease pool is still active.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>