Remove latest alias from docs
The latest alias was pointing to the latest release, including
pre-releases. Pre-releases, on the other hand, won't get their docs
published. This resulted in a situation in which the docs workflow was
broken as long as the latest version was a pre-release.

Fix this by removing the latest alias from the workflow. Also sync the
manual publish docs workflow with the automatic one.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit 2c5640c561c65d5e8e63990e49bbfae90b17627a)