Updated the `ha3x3` integration test to work with ap-embedded `k0s`
This updates the `ha3x3` integration test to use the new `forceupdate` feature
to force an update of the same `k0s` binary in test. The launch mode for this
integration test is using `openrc` which allows for the binary to be replaced.

To handle the difference in behaviour between ap-emebedded in `k0s` vs standalone,
the `restart` reconciler needed to be split into two pieces. This was necessary
as when ap was running standalone, all controller-runtime events appear as updates
(unless the process dies). When running embedded in `k0s`, our only indication that
the `k0s` binary has restarted is the 'created' event which is used when data is
loaded/cached (ie. on startup)

The previous `ha3x3` test would end by comparing the installed versions of `k0s`,
asserting them to whatever version is mandated in the integration test plan. Since
the same binary is used for testing (same version numbers), this has been expanded
to just assert that all state/status are completed across `controllers` and `workers`.

With this change in place, the remaining ap-integration tests can be ported over.

Signed-off-by: Shane Jarych <sjarych@mirantis.com>