Use a 'tools' image for all locally-installed tools
This introduces a tools image that contains all of the tools that are needed for building
k0s. Tools that have been moved to this image are: `client-gen`, `controller-gen`, `go-bindata`,
and `glangci-lint`.

Autopilot development used this image concept for its own `controller-gen` and `client-gen`
usages, however this has been adopted and extended for `k0s`.

Signed-off-by: Shane Jarych <sjarych@mirantis.com>