Move controller `autopilot` creation into a `Component`
All of the logic for creating + starting autopilot on Controllers was
inline with the controller initialization. This has been moved into
its own `autopilot.go`, similar to what was done in workers.

This did uncover + fix a bug where autopilot would not be started
if `--enable-worker` was provided. The current expectation for
controllers with `--enable-worker` is that the default autopilot
`controller` behavior will be used.

Signed-off-by: Shane Jarych <sjarych@mirantis.com>