Added a generic 'apply failed' status to plans + nodes.
Previously, when an update failed at the signal node level, the state in
the plan would stay at 'SignalSent', which implies that something is
processing, despite the Plan being technically "done"

This change adds a 'PlanApplyFailed' to the plan states, and a new
`SignalApplyFailed` to signal node states.

If any signal node reports a `SignalApplyFailed`, the plan will switch
to `PlanApplyFailed`, and the plan will terminate.

Signed-off-by: Shane Jarych <sjarych@mirantis.com>