Fix a `SIGTERM` shutdown issue in supervisor
This cleans up the creation/closure of a channel used to indicate
that the supervisor goroutine is 'done'. The `defer` in the for-loop
wasn't being invoked per-iteration (due to the goroutine function
not exiting).

Also, this `done` channel has been changed to notify on `close()`
instead of sending data into it.

Signed-off-by: Shane Jarych <sjarych@mirantis.com>