Added the generic 'apply failed' to `airgapupdate`
This was added initially to the `k0supdate` command, but not `airgapupdate`.
The repurcussion was that if an `airgapupdate` is listed in a plan and fails,
the state for `airgapupdate` will stay at 'SignalSent'

Now, if an `airgapupdate` fails, it will return as `SignalApplyFailed`, allowing
the plan to transition to `PlanApplyFailed`.

Signed-off-by: Shane Jarych <sjarych@mirantis.com>