Fix GitHub release workflow
The build of the k0s binary aims to not rely on the Go toolchain
installed on the build host. The toolchain installation was removed
from the GitHub workflow that builds the Windows binary under that
assumption in fe4f1948d.

However, that assumption does not hold, as there's still one invocation
of the go binary inside a Makefile shell function. Building k0s when
there's no go binary in the path doesn't break the build, but produces
a binary that panics at runtime, since some of the variables set at
link time have invalid values.

Fix this by determining the link time variables via another mechanism
that doesn't rely on the go binary. Add additional tests that check for
consistency between the binary versions of the embedded binaries with
their respective Go module dependencies, so that it's safe to assume
that they are in sync.

Move down the toolchain installation in the other release builds just
before the smoke tests, as they are the only ones which require it.

Copy over the ARM toolchain installation procedure from the PR workflow
to the release workflow, which also ensures that the ARM release
actually uses a fixed Go version to execute the smokes.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit 05c3c33edaaa5e4ecd7defa1939e1fa043aed98b)