Use a custom GitHub token for checkouts in the backport action
The default GitHub token is not entitled to make changes to GitHub
workflows. Hence, backporting changes to workflows will fail, since the
git push operation that touches the workflow files will be rejected with
a permission error:

> ! [remote rejected]   backport-1954-to-release-1.23 -> backport-1954-to-release-1.23 (refusing to allow a GitHub App to create or update workflow `.github/workflows/go.yml` without `workflows` permission)

Use the same token for git pull/push that is used for GitHub API access.
This one will have the appropriate permissions.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>