Cache airgap image bundle in all builds
The airgap image bundles are now also checked on ARM builds. This means
a lot of image pulls from docker.io, which is rate limited. Mitigate
this by adding a caching layer for airgap image bundles for the regular
amd64 and the custom armv7/arm64 builds.

Embed the TARGET_PLATFORM setting into the Makefile, add the platform
name to the file names and add some auto-detection of the architecture.
This way, the TARGET_PLATFORM doesn't need to be specified explicitly.
Calculate a cache key for the image bundle based on the hash of the
Makefile, the airgap image bundle list and the image-bundler Docker
image that is used to produce the bundles.

Move the image-bundler Docker image to the hack/ folder and remove the
necessity to embed the image list into the image or to mount folders
into the container. Let the bundler script fail in case of an error and
suppress the very noisy output of the ctr CLI tool.

Change the integration tests to only mount the airgap image bundle if
this is explicitly requested by an integration test. Support multiple
mount-points, too.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit 2d73409fe0fc9b95bfc9d5209f104aed39480748)