Add new workflow to build docs
Up until now, there was no workflow that was building the docs for Pull
Requests. Since automatic dependency updates are now enabled, it's quite
important to check if the docs can actually be built. This won't catch
any rendering errors, but will at least ensure that the process remains
functional.

Moreover: Add caching for the Python dependencies and don't upgrade pip
manually, as this is already done by the setup-python action.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>