Move Dockerfile args down to where they are used
This improves the layer caching a bit, i.e. the dependencies don't need
to be reinstalled when the binary versions change, or the upstream
sources don't need to be cloned when the build flags change.

Moreover: remove the CMD stanzas from the Dockerfiles. The images are
never meant to be run anyways. Just provide a dummy entrypoint in order
to create containers for exporting the binaries.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>