Use GNU Make 3.81 as baseline
Document the build requirements of k0s a bit and explicitly specify the
requirement of GNU Make at a minimum version of 3.81. This seems to be
the common baseline on vanilla macOS installs, so being compatible with
that version reduces a bit of friction when starting to hack on k0s on
macs.

Take care of some old links, too.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>