Add the output of k0s sysinfo to bug template
Oftentimes k0s version and operating system name and version aren't
enough context to judge about the problems that people report. Moreover,
the OS release doesn't include the architecture, so we cannot tell if
it's on arm or amd.

k0s already features `k0s sysinfo` that prints some diagnostics about
the system. Include this as a separate form field and provide a default
value that allows for collapsing the output. (Sadly, there's no way to
make a form field collapsible by default, so this is less convenient to
use as it should be.)

Also extend the suggested command to get the platform information by
also including uname (whose info should probably be included into k0s's
diagnostic commands.)

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>