Don't load cluster config automatically
The cluster-wide config is only required in a few cases, but is eagerly
loaded for every subcommand. This is unnecessary and leads to a
situation in which a controller started with --enable-dynamic-config
tries to connect to itself to load the config, hanging in a retry loop
for several minutes. The logging configuration to enable debug logging
is evaluated later than that retry loop, so it seems that the controller
is dead, with no log output whatsoever.

Remove the ClusterConfig field from the CLIOptions struct and make the
private loading function public. Use that function to load the config on
demand in places that require it.

Also replace the usage of ClusterConfig with NodeConfig in the etcd
command. The storage spec is a node-local configuration and shouldn't be
read from the cluster config in the first place.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>