Write manifests atomically
Use a temporary-write-and-rename approach to write manifests to disk, so
that the appliers never see partially written files. The strategy is as
follows: Create a temporary file in the same directory as the target,
write to it, sync and close the file handle, do a chmod and a rename. As
a result, there's no more need to ignore chmod events in the stack
applier. Those events actually *can* matter, when files become readable
or unreadable for the k0s process.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>