Use named return values in WriteAtomically
The intention was to use a named return value, so that the deferred
function would be able to modify it. Without the named value, the error
handling in the deferred function has no effect.

Moreover: Add a bunch of tests and suppress any errors about failed
removals of temporary files indicating that the files don't exist.

Fixes: 8ea562661 ("Write manifests atomically")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>