Ignore deprecation lints in CI
Deprecations shouldn't break the CI. If this were the case, it would
somehow be pointless in having them, as adding a deprecation would have
the same effect as introducing a breaking API change directly.

Albeit it would be nice to have some indication of deprecations being
introduced in CI builds, it's currently impossible to do so without
having a second linter run with a different config, whose non-zero error
code won't break the build.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>