Fix docs build after pip 22.3 update
The setup-python GitHub action automatically performs an update to the
latest version of pip. Pip 22.3 got released on 2022-10-15, which had a
deprecation that broke the k0s docs build. Add an additional
requirements file that installs pip and wheel as build prerequisites, so
that the docs build can control the exact version of pip in use, as this
seems to be something that influences the build stability.

Also add a separate build that tests whether the Docker container for
local docs development can be built.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>