Added 'toolsuite' with a plan-apply operation
This introduces a new test suite implementation that will use 'tool' to create
a `k0s` cluster on setup, and destroy it on tear-down. In between, custom operations
are applied that perform tests on the cluster.

The initial operation will build an autopilot 'Plan' and apply it. The determination
of success is the test observing that the Plan transitions to 'Completed'

Signed-off-by: Shane Jarych <sjarych@mirantis.com>