Don't suppress retries for server errors in watch
During Kubernetes startup it may happen that the API server is not yet
fully initialized and answers with 503 and a message of "the request has
been made before all known HTTP paths have been installed, please try
again".

It seems reasonable to not forcefully bypass the retry logic for all
server errors that specify a retry-after hint.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>