Add context to ToolSuite in inttests
Prepare ToolSuite-based tests for test cancellation. Add a Context
method to the suite struct that may be used to obtain some context
whenever this is necessary during testing. This mimics the functionality
of FootlooseSuite-based tests.

The context returned by the new method is currently just context.TODO(),
but this can be changed with minimal impact on the rest of the tests
whenever actual test cancellation is required.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>