Unify make prerequisites and gen-bindata args for static assets
Introduce static_asset_dirs make variable and use that to drive both the
make prerequisites and the gen-bindata inputs. This ensures that those
are always in sync (the misc folder was missing before this). It
addresses the issue fixed in 92890e96d, too. So it's not required
anymore to delete the output file before calling gen-bindata, but it
also won't harm.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>