Added new GHA workflow for autopilot version updates
This workflow leverages both `toolsuite` and `tool` to create a specialized
AWS VPC, and performs autopilot updates of k0s releases.

Currently, this is intentionally a manual workflow due to the large number of
AWS resources created.

The current strategy of automated release testing is:

 - all branch releases should update to the latest branch release (ie. v1.24.3+k0s.0 -> v1.24.6+k0s.0)
 - the latest branch release should update to the latest main release (ie. v1.24.6+k0s.0 -> v1.25.2+k0s.0)

In addition, this change includes:

 - bugfix: workaround for v1.24 when `images = null`
 - clarified the concept of 'cluster configuration' instead of overloading the term 'kubeconfig'
 - allow tool to run without a TTY
 - bugfix: workaround for YAML persisting autopilot plans without `status`

Signed-off-by: Shane Jarych <sjarych@mirantis.com>