Let the ARM based smoke tests not be fail-fast
This is done for the other integration tests, too. The test failures are
almost always specific to certain jobs, and it's counterproductive to
cancel other jobs that are still running, but are likely to succeed.
Those cancelled jobs need to be re-run along with the failed tests, even
this wouldn't have been necessary if they hadn't been cancelled in the
first place.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>