Introduce Normalize method for controller options
Move the existing validation out of controller.go into a method on
ControllerOptions, so that the functionality becomes easier to test.
Moving the logic also allows for replacing AvailableCommands() with a
private global variable availableCommands, since it's no longer required
to be visible outside the config package.

Also: Add missing components to the list of available components:

* api-config
* autopilot
* node-role

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>