Fix the test custom-domain flakyness
The test custom-domain has become flaky. I'm not quite sure when did
this happen, but it was recently. We can see in many executions such as:
https://github.com/k0sproject/k0s/actions/runs/3742332086/jobs/6359020341
https://github.com/k0sproject/k0s/actions/runs/3740780796/jobs/6349616080
https://github.com/k0sproject/k0s/actions/runs/3731323166/attempts/2

We can solve this simply by waiting for pod logs as this guarantees that
konnectivity is actually working. It's worth noting that if there were
more controllers waiting for pod logs wouldn't be 100% reliable.

Signed-off-by: Juan Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>