Fix supervisor stop while respawn unit test on Windows
On Windows, sending the null signal to a process will always return an
error, so the test wasn't actually ensuring that the Supervisor was in
the unit test's desired state. Check the returned error and add an
additional sleep time to mitigate that. Moreover, to make the test more
portable, add some fallbacks for the command being used.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>