Fix gen-bindata unit test on Windows
On Windows, files and directories which are in use cannot be deleted.
The test switches the process working directory to a temporary folder
managed by the go testing framework. That temporary dir will be
attempted to be deleted after the test, which fails, since the process
still uses it as its working directory.

Fix that by introducing the new testutil.Chdir helper that can be
deferred to switch back to the original working directory.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>