Improve CoreDNS affinity check in basic inttest
The inttest sometimes fails with the following message:

    basic_test.go:217:
        	Error Trace:	/home/gh-runner/actions-runner/_work/k0s/k0s/inttest/basic/basic_test.go:217
        	            				/home/gh-runner/actions-runner/_work/k0s/k0s/inttest/basic/basic_test.go:105
        	Error:      	Should not be: ""
        	Test:       	TestBasicSuite/TestK0sGetsUp

This indicates that the test doesn't wait until the CoreDNS Pods are properly
scheduled. Maybe because it simply checks too early, or because the cluster is
not healthy.

Replace the one-shot affinity check with a watch on the Pods that waits until
two of them get scheduled. Add some logging and a better error message.

Moreover: Add label selector functionality to the Kubernetes
Watcher in a similar manner as the field selectors are
implemented. Remove WithoutFieldSelector() as it can be replaced by
WithFieldSelector(fields.Everything()). Leave a note about this in the docs.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>