Fix docs Python Docker image
The CI didn't trigger the docs build on changes to embedded-bins/
Makefile.variables. This lead to the effect that CI didn't catch a non-
existing Python Docker image.

Update to the current point release of Python, so that the correct
Docker image is available again. Also trigger the docs CI workflows on
changes to the aforementioned Makefile.

Fixes: d8bc98cc4 ("Bump Alpine to 3.17.1")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>