Introduce WaitForPlanState helper in AP inttests
Many inttests used the WaitForPlanByName helper to wait for a certain
terminal Plan state. When the Plan reached another terminal state than
the expected one, the tests would just hang and fail with a timeout.

Improve fail-fastness and error reporting for those cases by introducing
the aforementioned helper that'll stop watching for changes as soon as a
terminal state has been reached, returning an error in those cases where
there was a state mismatch.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>