Fix copyright in the codegen
The script validates the copyright of every go file that we own as a
codegen target. I think the most logical place to do this is in the make
lint, however auto generated files inject a copyright from a static
file, so we need to modify the auto generated code before the linting.

So I see two options:
1- Verify everything in the codegen
2- Replace the autogenerated parts in the codegen and verify everything
in the lint.

Since the lint depends on the codegen, I decided that the simplicity of
having just one script with just one call was more worthy than the
correctness.

Co-authored-by: Tom Wieczorek <twz123@users.noreply.github.com>
Signed-off-by: Juan Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>
(cherry picked from commit b78cc097c026a8b8202f427f4947d92b6bbbd198)