Detangle autopilot waitfor code
The code in waitfor.go was mainly used within the integration tests.
This is "okay", but it actually pulled in the retry logic which was
dedicated to the integration tests, which is - at least - unexpected.

Move the waitfor code into the inttest folder and inline the parts that
are actually required by autopilot production code, accounting for a
proper "production" retry logic, performing some refactorings on the
way:

* Check the "Established" condition when waiting for CRDs. This is the
  proper way of ensuring that a CRD is actually usable.
* Remove some generic helpers and the condition varargs parameters, as
  they were no longer required after detangling everything and a bit of
  copy/paste is okay.
* Remove the timeout parameters from the inttest functions. There's
  already the inttest context in place which effectively interrupts the
  test execution when the test times out.
* Remove unnecessary return values after inlining the generic code.
* Add the autopilot API group automatically, so this isn't duplicated
  at all the call sites.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>