Integrate Troubleshoot support bundle collection into inttests
See https://troubleshoot.sh/. This replaces kubectl cluster-info --dump.
An advantage of the support bundle is that it can be used in conjunction
with sbctl to spin up a fake API server that reflects the cluster's
state at the time of the support bundle creation. This makes it much
easier to inspect than having to deal with all the individual JSON
files using find/grep/jq. See https://github.com/replicatedhq/sbctl for
details on that.

Moreover: Combine all upload CI steps into a single one. No need to have
them separately. Only collect the logs if the tests failed. The CI step
is also just triggered on failure.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>