Fix airgap bundle caching in CI
The current implementation assumed that the evaluation of the hashFiles
function would happen at script evaluation time, just as every other
script command. That assumption is false, as the hashing functionality
is not part of the script, but part of the template engine that creates
the script's content. That means that all files which should take part
in hashing need to be in place during step creation time.

The broken hash calculation lead to the usage of wrong image bundles
during CI, since the airgap-images.txt file did not change the hash in
any way. Fix that by first generating the image list, and calculate the
hash in a separate subsequent build step.

Fixes: 2d73409fe ("Cache airgap image bundle in all builds")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit 5df21bdcf7afc755125c56ad3aa461d9d3b3b179)