Fix external address reconciliation w.r.t. dynamic config
The external address is not part of the reconcilable cluster config, but
part of the non-reconcilable node config. The external API endpoint
reconciler was looking at the cluster config. This is fine as long as
dynamic config isn't enabled, because the config splitting doesn't
happen then. When dynamic config is enabled, the splitting takes place
and the reconciler tries to reconcile an empty external address,
resulting in log statements like these:

    level=error msg="cannot resolve api.externalAddress: lookup : no such host" component=endpointreconciler
    level=warning msg="external API address reconciliation failed: lookup : no such host" component=endpointreconciler

Fix this by removing the cluster config reconciliation from the
reconciler and use the node config's API address instead.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>