Take kube-proxy API address from node config
The API spec is not part of the dynamic cluster config. Taking the
API address from there will not work properly when dynamic config is
enalbed, e.g. when there are port overrides for the API server.

This is already fixed in 1.26 and onwards, but there's no simple commit
to be backported. Hence create a new minimal patch for 1.25.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit ef6e804e31db956f23a41391e2a1621ec29619b8)