Fix k0s status components
If a component.Healthy returns an error which is not nil the json would
be marshalled and unmarshalled incorrectly. This happened because error
is an interface which doesn't implement json.Marshaller and
json.Unmarshalled interfaces.

The new type ProbeError is internally a string which can be marshalled
and unmarshalled and still allows users to use errors in a standard way.

Co-authored-by: Kimmo Lehto <kimmo.lehto@gmail.com>
Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>