Better error reporting in CSRApprover
Refactor the CSRApprover component in a way that errors are reported in
a more structured and detailed way. Let all bool functions return errors
instead, so that they can explain what was rejected. Improve the logging
of those errors, so that they aren't logged twice and the CSR names are
included in the logs. Remove the csrRecognizer on the way, as there is
only a single one, and there's no additional ones on the horizon, so
this seemed unnecessary, and inlining it makes the approval loop easier
to read.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>